Add Animated sprite (#57)
* Change Sprite and Display to use uint16_t for dimensions
* Add PGMFile and gifdec

+++ src/Display/Display.h
@@ -18,7 +18,7 @@ public:
 	 * @param rotation
 	 * @param mirror Horizontal display mirroring. WARNING: decimates the framerate and uses width*height*2 bytes of RAM
 	 */
-	Display(uint8_t width, uint8_t height, int8_t blPin = -1, int8_t rotation = -1, bool mirror = false);
+	Display(uint16_t width, uint16_t height, int8_t blPin = -1, int8_t rotation = -1, bool mirror = false);
 	void begin();
 	void commit();
 	void clear(uint32_t color);
@@ -38,8 +38,8 @@ public:
 
 private:
 	TFT_eSPI tft; 
-	uint width;
-	uint height;
+	uint16_t width;
+	uint16_t height;
 	int8_t blPin;
 	int8_t rotation;
